Added support for multiple keys as per i18n documentation: ``` i18n.t(["notExists", "key1"]); // -> value of key 1 ```
To keep existing features and have that available in templates through helpers, I added a `keys` option. This option is only used if you don't specify a key to translate:
```
{{_ keys='tasks.new,new' context='female'}}
```

This will try to find a key named `tasks.new` and, if not found, will try to find a key named `new`. It will still pass the context option as usual.